Added ReadMe for Prefab and Changed Enemy stats

+++ WonkyWizards/Assets/src/david/Scripts/EnemyTypes/GoblinGiant.cs
@@ -138,7 +138,7 @@ public class GoblinGiant : Enemy
             agent.acceleration += speed_amount;
         }
     }
-    public static float GetDeathDamage() 
+    public static int GetDeathDamage() 
     {
         return death_damage;
     }
